gom: Fix a format warning
By casting GTypes to ints to print them.

+++ gom/gom-command.c
@@ -59,7 +59,7 @@ gom_command_bind_param (GomCommand   *command,
 
    priv = command->priv;
 
-   g_debug("Binding gtype %s (%d).", g_type_name(G_VALUE_TYPE(value)), G_VALUE_TYPE(value));
+   g_debug("Binding gtype %s (%d).", g_type_name(G_VALUE_TYPE(value)), (int) G_VALUE_TYPE(value));
 
    switch (G_VALUE_TYPE(value)) {
    case G_TYPE_BOOLEAN:
@@ -151,7 +151,7 @@ gom_command_bind_param (GomCommand   *command,
          }
          break;
       }
-      g_warning("Failed to bind gtype %s (%d).", g_type_name(G_VALUE_TYPE(value)), G_VALUE_TYPE(value));
+      g_warning("Failed to bind gtype %s (%d).", g_type_name(G_VALUE_TYPE(value)), (int) G_VALUE_TYPE(value));
       g_assert_not_reached();
       break;
    }
